Backend/Server/src/requirements.txt
astroid==2.15.6
blinker==1.6.2
click==8.1.7
dill==0.3.7
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
greenlet==2.0.2
gunicorn==21.2.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.3
mccabe==0.7.0
packaging==23.1
platformdirs==3.10.0
pylint==2.17.5
python-dotenv==1.0.0
SQLAlchemy==2.0.21
tomlkit==0.12.1
typing_extensions==4.8.0
Werkzeug==2.3.7
wrapt==1.15.0
